``` Refactor Livewire components into organized namespaces
- Moved Dashboard to Livewire\Assets namespace
- Moved Auth and Logout to Livewire\Users namespace
- Updated references and routes accordingly
- Added Locked attribute and email_otp property for local OTP handling
```
